Break match results data out from main CandyStore class
